requirements.txt
streamlit>=1.36
pandas>=2.0
numpy>=1.26
plotly>=5.20
scikit-learn>=1.3
umap-learn>=0.5.6
sentence-transformers>=3.0
